+++ b3j0f/conf/driver/file/json.py
@@ -70,7 +70,7 @@ class JSONConfDriver(FileConfDriver):
 
         return result
 
-    def _set_conf(self, conf, resource, rscpath):
+    def _setconf(self, conf, resource, rscpath):
 
         for category in conf: